pyproject.toml
[tool.poetry]
name = "alignment"
version = "0.1.0"
description = ""
authors = ["Caroline Binley"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
spacy = "^3.7.5"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz"}
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
openai = "^1.40.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
notebook = "^7.2.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
